PIM-4338: Clean DatagridBundle
