+++ src/Pim/Bundle/DataGridBundle/DependencyInjection/Compiler/AddAttributeTypesPass.php
@@ -15,7 +15,7 @@ use Symfony\Component\DependencyInjection\ContainerBuilder;
 class AddAttributeTypesPass implements CompilerPassInterface
 {
     /** @staticvar string */
-    const REGISTRY_ID = 'pim_datagrid.datagrid.product.configuration_registry';
+    const REGISTRY_ID = 'pim_datagrid.datagrid.configuration.product.configuration_registry';
 
     /** @staticvar string */
     const PARAM_PREFIX = 'pim_datagrid.product.attribute_type.';
