+++ src/Pim/Bundle/DataGridBundle/Resources/config/configurators.yml
@@ -1,15 +1,14 @@
 parameters:
-    pim_datagrid.datagrid.product.context_configurator.class: Pim\Bundle\DataGridBundle\Datagrid\Product\ContextConfigurator
-    pim_datagrid.datagrid.product.columns_configurator.class: Pim\Bundle\DataGridBundle\Datagrid\Product\ColumnsConfigurator
-    pim_datagrid.datagrid.product.filters_configurator.class: Pim\Bundle\DataGridBundle\Datagrid\Product\FiltersConfigurator
-    pim_datagrid.datagrid.product.sorters_configurator.class: Pim\Bundle\DataGridBundle\Datagrid\Product\SortersConfigurator
-    pim_datagrid.datagrid.product.group_columns_configurator.class: Pim\Bundle\DataGridBundle\Datagrid\Product\GroupColumnsConfigurator
-    pim_datagrid.datagrid.request_parameters_extractor.class: Pim\Bundle\DataGridBundle\Datagrid\RequestParametersExtractor
+    pim_datagrid.datagrid.configuration.product.context_configurator.class:       Pim\Bundle\DataGridBundle\Datagrid\Configuration\Product\ContextConfigurator
+    pim_datagrid.datagrid.configuration.product.columns_configurator.class:       Pim\Bundle\DataGridBundle\Datagrid\Configuration\Product\ColumnsConfigurator
+    pim_datagrid.datagrid.configuration.product.filters_configurator.class:       Pim\Bundle\DataGridBundle\Datagrid\Configuration\Product\FiltersConfigurator
+    pim_datagrid.datagrid.configuration.product.sorters_configurator.class:       Pim\Bundle\DataGridBundle\Datagrid\Configuration\Product\SortersConfigurator
+    pim_datagrid.datagrid.configuration.product.group_columns_configurator.class: Pim\Bundle\DataGridBundle\Datagrid\Configuration\Product\GroupColumnsConfigurator
+    pim_datagrid.datagrid.request.parameters_extractor.class:                     Pim\Bundle\DataGridBundle\Datagrid\Request\RequestParametersExtractor
 
 services:
-
-    pim_datagrid.datagrid.product.context_configurator:
-        class: %pim_datagrid.datagrid.product.context_configurator.class%
+    pim_datagrid.datagrid.configuration.product.context_configurator:
+        class: %pim_datagrid.datagrid.configuration.product.context_configurator.class%
         arguments:
             - '@pim_catalog.manager.product'
             - '@oro_datagrid.datagrid.request_params'
@@ -18,33 +17,57 @@ services:
         calls:
             - [setRequest, ['@?request=']]
 
-    pim_datagrid.datagrid.product.columns_configurator:
-        class: %pim_datagrid.datagrid.product.columns_configurator.class%
+    pim_datagrid.datagrid.configuration.product.columns_configurator:
+        class: %pim_datagrid.datagrid.configuration.product.columns_configurator.class%
         arguments:
-            - '@pim_datagrid.datagrid.product.configuration_registry'
+            - '@pim_datagrid.datagrid.configuration.product.configuration_registry'
 
-    pim_datagrid.datagrid.product.filters_configurator:
-        class: %pim_datagrid.datagrid.product.filters_configurator.class%
+    pim_datagrid.datagrid.configuration.product.filters_configurator:
+        class: %pim_datagrid.datagrid.configuration.product.filters_configurator.class%
         arguments:
-            - '@pim_datagrid.datagrid.product.configuration_registry'
+            - '@pim_datagrid.datagrid.configuration.product.configuration_registry'
 
-    pim_datagrid.datagrid.product.sorters_configurator:
-        class: %pim_datagrid.datagrid.product.sorters_configurator.class%
+    pim_datagrid.datagrid.configuration.product.sorters_configurator:
+        class: %pim_datagrid.datagrid.configuration.product.sorters_configurator.class%
         arguments:
-            - '@pim_datagrid.datagrid.product.configuration_registry'
+            - '@pim_datagrid.datagrid.configuration.product.configuration_registry'
 
-    pim_datagrid.datagrid.product.group_columns_configurator:
-        class: %pim_datagrid.datagrid.product.group_columns_configurator.class%
+    pim_datagrid.datagrid.configuration.product.group_columns_configurator:
+        class: %pim_datagrid.datagrid.configuration.product.group_columns_configurator.class%
         arguments:
-            - '@pim_datagrid.datagrid.product.configuration_registry'
+            - '@pim_datagrid.datagrid.configuration.product.configuration_registry'
             - '@oro_datagrid.datagrid.request_params'
             - '@pim_catalog.repository.group'
         calls:
             - [setRequest, ['@?request=']]
 
-    pim_datagrid.datagrid.request_parameters_extractor:
-        class: %pim_datagrid.datagrid.request_parameters_extractor.class%
+    pim_datagrid.datagrid.request.parameters_extractor:
+        class: %pim_datagrid.datagrid.request.parameters_extractor.class%
         arguments:
             - '@oro_datagrid.datagrid.request_params'
         calls:
             - [setRequest, ['@?request=']]
+
+    # deprecated, will be removed in 1.5, please use pim_datagrid.datagrid.configuration.product.context_configurator
+    pim_datagrid.datagrid.product.context_configurator:
+        alias: pim_datagrid.datagrid.configuration.product.context_configurator
+
+    # deprecated, will be removed in 1.5, please use pim_datagrid.datagrid.configuration.product.columns_configurator
+    pim_datagrid.datagrid.product.columns_configurator:
+        alias: pim_datagrid.datagrid.configuration.product.columns_configurator
+
+    # deprecated, will be removed in 1.5, please use pim_datagrid.datagrid.configuration.product.filters_configurator
+    pim_datagrid.datagrid.product.filters_configurator:
+        alias: pim_datagrid.datagrid.configuration.product.filters_configurator
+
+    # deprecated, will be removed in 1.5, please use pim_datagrid.datagrid.configuration.product.sorters_configurator
+    pim_datagrid.datagrid.product.sorters_configurator:
+        alias: pim_datagrid.datagrid.configuration.product.sorters_configurator
+
+    # deprecated, will be removed in 1.5, please use pim_datagrid.datagrid.configuration.product.group_columns_configurator
+    pim_datagrid.datagrid.product.group_columns_configurator:
+        alias: pim_datagrid.datagrid.configuration.product.group_columns_configurator
+
+    # deprecated, will be removed in 1.5, please use pim_datagrid.datagrid.request.parameters_extractor
+    pim_datagrid.datagrid.request_parameters_extractor:
+        alias: pim_datagrid.datagrid.request.parameters_extractor
