+++ src/Pim/Bundle/DataGridBundle/Resources/config/registry.yml
@@ -1,6 +1,10 @@
 parameters:
-    pim_datagrid.datagrid.product.configuration_registry.class: Pim\Bundle\DataGridBundle\Datagrid\Product\ConfigurationRegistry
+    pim_datagrid.datagrid.configuration.product.configuration_registry.class: Pim\Bundle\DataGridBundle\Datagrid\Configuration\Product\ConfigurationRegistry
 
 services:
+    pim_datagrid.datagrid.configuration.product.configuration_registry:
+        class: %pim_datagrid.datagrid.configuration.product.configuration_registry.class%
+
+    # deprecated, will be removed in 1.5, please use pim_datagrid.datagrid.configuration.product.configuration_registry
     pim_datagrid.datagrid.product.configuration_registry:
-        class: %pim_datagrid.datagrid.product.configuration_registry.class%
+        alias: pim_datagrid.datagrid.configuration.product.configuration_registry
