+++ src/Pim/Bundle/EnrichBundle/Resources/config/datagrid_listeners.yml
@@ -9,10 +9,10 @@ services:
     pim_enrich.event_listener.product_grid_before_listener:
         class: %pim_datagrid.event_listener.configure_product_grid_listener.class%
         arguments:
-            - '@pim_datagrid.datagrid.product.context_configurator'
-            - '@pim_datagrid.datagrid.product.columns_configurator'
-            - '@pim_datagrid.datagrid.product.filters_configurator'
-            - '@pim_datagrid.datagrid.product.sorters_configurator'
+            - '@pim_datagrid.datagrid.configuration.product.context_configurator'
+            - '@pim_datagrid.datagrid.configuration.product.columns_configurator'
+            - '@pim_datagrid.datagrid.configuration.product.filters_configurator'
+            - '@pim_datagrid.datagrid.configuration.product.sorters_configurator'
         tags:
             - { name: kernel.event_listener, event: oro_datagrid.datgrid.build.before.product-grid, method: buildBefore }
 
@@ -31,10 +31,10 @@ services:
     pim_enrich.event_listener.product_group_grid_before_listener:
         class: %pim_datagrid.event_listener.configure_product_grid_listener.class%
         arguments:
-            - '@pim_datagrid.datagrid.product.context_configurator'
-            - '@pim_datagrid.datagrid.product.group_columns_configurator'
-            - '@pim_datagrid.datagrid.product.filters_configurator'
-            - '@pim_datagrid.datagrid.product.sorters_configurator'
+            - '@pim_datagrid.datagrid.configuration.product.context_configurator'
+            - '@pim_datagrid.datagrid.configuration.product.group_columns_configurator'
+            - '@pim_datagrid.datagrid.configuration.product.filters_configurator'
+            - '@pim_datagrid.datagrid.configuration.product.sorters_configurator'
         tags:
             - { name: kernel.event_listener, event: oro_datagrid.datgrid.build.before.product-group-grid, method: buildBefore }
 
@@ -54,10 +54,10 @@ services:
     pim_enrich.event_listener.product_variant_group_grid_before_listener:
         class: %pim_datagrid.event_listener.configure_product_grid_listener.class%
         arguments:
-            - '@pim_datagrid.datagrid.product.context_configurator'
-            - '@pim_datagrid.datagrid.product.group_columns_configurator'
-            - '@pim_datagrid.datagrid.product.filters_configurator'
-            - '@pim_datagrid.datagrid.product.sorters_configurator'
+            - '@pim_datagrid.datagrid.configuration.product.context_configurator'
+            - '@pim_datagrid.datagrid.configuration.product.group_columns_configurator'
+            - '@pim_datagrid.datagrid.configuration.product.filters_configurator'
+            - '@pim_datagrid.datagrid.configuration.product.sorters_configurator'
         tags:
             - { name: kernel.event_listener, event: oro_datagrid.datgrid.build.before.product-variant-group-grid, method: buildBefore }
 
@@ -94,10 +94,10 @@ services:
     pim_enrich.event_listener.association_product_grid_before_listener:
         class: %pim_datagrid.event_listener.configure_product_grid_listener.class%
         arguments:
-            - '@pim_datagrid.datagrid.product.context_configurator'
-            - '@pim_datagrid.datagrid.product.columns_configurator'
-            - '@pim_datagrid.datagrid.product.filters_configurator'
-            - '@pim_datagrid.datagrid.product.sorters_configurator'
+            - '@pim_datagrid.datagrid.configuration.product.context_configurator'
+            - '@pim_datagrid.datagrid.configuration.product.columns_configurator'
+            - '@pim_datagrid.datagrid.configuration.product.filters_configurator'
+            - '@pim_datagrid.datagrid.configuration.product.sorters_configurator'
         tags:
             - { name: kernel.event_listener, event: oro_datagrid.datgrid.build.before.association-product-grid, method: buildBefore }
