+++ src/Pim/Bundle/FilterBundle/Resources/config/filters.yml
@@ -97,7 +97,7 @@ services:
         arguments:
             - '@form.factory'
             - '@pim_filter.product_utility'
-            - '@pim_datagrid.datagrid.request_parameters_extractor'
+            - '@pim_datagrid.datagrid.request.parameters_extractor'
         tags:
             - { name: oro_filter.extension.orm_filter.filter, type: product_in_group }
 
@@ -106,7 +106,7 @@ services:
         arguments:
             - '@form.factory'
             - '@pim_filter.product_utility'
-            - '@pim_datagrid.datagrid.request_parameters_extractor'
+            - '@pim_datagrid.datagrid.request.parameters_extractor'
             - '@pim_catalog.repository.association_type'
             - '@pim_catalog.manager.product'
         tags:
